Use Sidekiq wrapped option for better logging
When using a wrapper class, Sidekiq's logging will show that class name
for each job that's performed. So, using sidekiq-clutch logs a lot of
Sidekiq::Clutch::JobWrapper jobs. To make the logs a bit more useful, we
can pass the underlying job class name as the `wrapped` option, and
Sidekiq will use that for the logs instead.