Treat each parallel block as a distinct step
Before this change, consecutive parallel blocks would effectively be
treated (and ultimately processed) as a single step of work to be done
in parallel. For example:

    clutch.parallel do
      widgets.each do |w|
        clutch.jobs << [Step1, w.id]
      end
    end
    clutch.parallel do
      widgets.each do |w|
        clutch.jobs << [Step2, w.id]
      end
    end

This code, which might aims to perform Step1 on each widget, then
perform Step2 on each widget when Step1 has finished, would actually
perform Step1 and Step2 at the same time.

This change treats each parallel block as a distinct step.